feat: add filters

+++ Million.BackEnd/Million.BackEnd.Contracts/Common/Responses.cs
@@ -1,4 +1,4 @@
 ﻿namespace Million.BackEnd.Contracts.Common
 {
-    public record PaginationResponse<T>(int Total, T Page);
+    public record PaginationResponse<T>(long Total, T Page);
 }
